Make S3 functionality more test-friendly
- Allow overriding s3_client
- Expose LoginGov::Hostdata::FakeS3Client for use in tests
