+++ lib/login_gov/hostdata/version.rb
@@ -1,5 +1,5 @@
 module LoginGov
   module Hostdata
-    VERSION = "0.3.1"
+    VERSION = "0.3.2"
   end
 end
